requirements.txt
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.13
albumentations>=1.3.0
opencv-python>=4.5.5.64
facenet-pytorch>=2.5.2
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.1.0
tqdm>=4.64.0
tensorboard>=2.11.0
Pillow>=9.0.0
matplotlib>=3.7.0
pyyaml>=6.0
seaborn>=0.12.0